mas avances

+++ entregaFinal-Marrero/src/services/product.service.js
@@ -1,23 +1,33 @@
-import  { productModel } from "../db/models/productModel.js";
+import  { productModel } from "../db/models/product.model.js";
+import mongoose from 'mongoose';
 
-class ProductService {
+export class ProductService {
 
     
-    async getAll() { //TODO : agregar filter y demas
-        return this.products;
+    async getAll() {
+        const allProduct = await productModel.find();
+        return allProduct;
     }
 
 
     async getById(id) {
        try {
-        const product = await productModel.findOne({_id: id});
-
-        if (!product) {
+        
+        if (!mongoose.Types.ObjectId.isValid(id)) {
             return null;
         }
 
+        const product = await productModel.findById(id);
+
+        if (!product) {
+            null;
+        }
+
+        return product
+
         } catch (error) {
-            console.error(`An error occurred while try to get the product: ${id} with error message: ${error.message}`);
+            console.error(`Error occurred while trying to get the product with ID: ${id} - ${error.message}`);
+            return error;        
         }
     }
 
@@ -34,13 +44,15 @@ class ProductService {
     }){
 
         try{
+
             const newProduct = productModel.create({title, description, code, price, status, stock, category, thumbnail});  
             return newProduct;
 
         }
         catch(error) {
-
             console.error(`An error occurred while try to create a product with error message: ${error.message}`);
+            return error;        
+
         }
     }
 
@@ -57,6 +69,11 @@ class ProductService {
     }){
     
         try{
+            if (!mongoose.Types.ObjectId.isValid(id)) {
+                return null;
+            }
+
+            
             const updatedProduct = await productModel.findByIdAndUpdate(
                 id,
                 {
@@ -73,25 +90,39 @@ class ProductService {
                 },
                 { new: true, omitUndefined: true }  
             )
-            
+
+            if (!updatedProduct) {
+                null;
+            }
+    
             return updatedProduct;
 
         }catch(error) {
             console.error(`An error occurred while try to update the product: ${id} with error message: ${error.message}`);
-            return null;
+            return error;        
         }
     }
 
     async delete( id ) {
 
         try{
+
+            if (!mongoose.Types.ObjectId.isValid(id)) {
+                return null;
+            }
+
+            
             const deletedProduct = await productModel.deleteOne({_id: id});
+
+            if (!deletedProduct) {
+                null;
+            }
+
             return deletedProduct;
         }catch(error) {
             console.error(`An error occurred while try to delete the product: ${id} with error message: ${error.message}`);
-            return null;
+            return error;        
         }
     }
 }
 
-export { ProductService }
